django/requirements.txt
# primary file to manage dependency
# pipfile is obsoleted
requests==2.21.0
botocore==1.12.115
dj-database-url==0.5.0
django==2.2.13
django-cors-headers==3.0.2
django-filter==2.2.0
djangorestframework==3.10.0
docutils==0.14
gunicorn==19.9.0
jmespath==0.9.4
pyasn1==0.4.5
python-dateutil==2.8.0
pytz==2019.1
pyyaml==5.1
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
setuptools==41.6.0
coverage==4.5.4
nose==1.3.7
django-redis==4.12.1
psycopg2==2.8.6